# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  handlers.userResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about films
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT Token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/models.ApiError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Authenticate user
      tags:
      - auth
...
  /auth/user:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: User info
          schema:
            $ref: '#/definitions/handlers.userResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Get authenticated user info
      tags:
      - auth
  /genres:
    get:
      description: Retrieves a list of all available genres
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Genre'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get all genres
      tags:
      - genres
    post:
      consumes:
      - application/json
      description: Adds a new genre to the database
      parameters:
      - description: Genre data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: Created genre ID
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Couldn't bind json
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Create a new genre
      tags:
      - genres
  /genres/{id}:
    delete:
      description: Deletes a genre by ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Invalid genre id
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Delete a genre
      tags:
      - genres
    get:
      description: Retrieves a single genre by its ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Genre'
        "400":
          description: Invalid genre id
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get a genre by ID
      tags:
      - genres
    put:
      consumes:
      - application/json
      description: Updates an existing genre by ID
      parameters:
      - description: Genre ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated genre data
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/models.Genre'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid genre id or couldn't bind json
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Genre not found
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Update a genre
      tags:
      - genres
  /images/{imageId}:
    get:
      description: Retrieves an image file by its ID
      parameters:
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Binary image file
          schema:
            type: file
        "400":
          description: Invalid image id
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get image by ID
      tags:
      - images
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"